+++ apps/gesha-verify/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gesha-test"
+name = "gesha-verify"
 version = "0.1.0"
 edition = "2021"
