+++ apps/gesha-verify/src/main.rs
@@ -1,23 +1,23 @@
-use crate::test::Args;
+use crate::verify::Args;
 use clap::Parser;
 use gesha_core::{trace, Result};
 use std::process::ExitCode;
 use tracing::{error, info};
 
 mod overwrite;
-mod test;
+mod verify;
 
 #[tokio::main]
 async fn main() -> ExitCode {
     trace::init();
 
     let args = Args::parse();
-    info!("gesha-test: {:#?}", args);
+    info!("gesha-verify: {:#?}", args);
 
     let result = if args.overwrite {
         overwrite::run(args).await
     } else {
-        test::run(args).await
+        verify::run(args).await
     };
     trace::wait_to_export().await;
     to_code(result)
@@ -26,7 +26,7 @@ async fn main() -> ExitCode {
 fn to_code(result: Result<()>) -> ExitCode {
     match result {
         Ok(_) => {
-            info!("gesha-test: done");
+            info!("gesha-verify: done");
             ExitCode::SUCCESS
         }
         Err(cause) => {
