+++ apps/gesha-verify/src/overwrite.rs
@@ -1,20 +1,10 @@
-use clap::Parser;
+use crate::verify::Args;
 use gesha_core::conversions::{TestDefinition, TestRunner};
 use gesha_core::Result;
 use gesha_rust_shapes::v3_0;
-use tracing::instrument;
+use tracing::{info, instrument};
 
-#[derive(Parser, Debug)]
-#[clap(author, version, about, long_about = None)]
-pub struct Args {
-    #[clap(long)]
-    pub schema: Option<String>,
-
-    #[arg(long)]
-    pub overwrite: bool,
-}
-
-#[instrument(name = "test::run")]
+#[instrument(name = "overwrite::run")]
 pub async fn run(args: Args) -> Result<()> {
     let converter = v3_0::Converter::default();
     process(converter, args).await
@@ -26,6 +16,13 @@ async fn process<A: TestDefinition>(definition: A, args: Args) -> Result<()> {
     } else {
         definition.list_test_cases()
     };
+    let test_suites = definition.test_suites();
     let runner = TestRunner::new(definition);
-    runner.run_tests(cases).await
+    let modified_cases = runner.collect_modified_cases(cases).await?;
+    if modified_cases.is_empty() {
+        info!("diff not detected");
+    } else {
+        runner.copy_modified_files(&modified_cases)?;
+    }
+    runner.generate_test_suite_files(&test_suites)
 }
